separating logic to try and utilise server components as much as possible
