+++ src/app/income/page.tsx
@@ -1,42 +1,13 @@
-import { useState } from "react";
-import { useDispatch } from "react-redux";
-import { addIncome } from "../../redux/slices/incomeSlice";
+import { GetServerSideProps } from "next/types";
+import IncomeForm from "../income/IncomeForm";
+import IncomeList from "../income/IncomeList";
 
-export default function IncomePage() {
-  const [amount, setAmount] = useState("");
-  const [date, setDate] = useState("");
-  const [type, setType] = useState("");
-  const dispatch = useDispatch();
-
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    dispatch(addIncome({ amount: parseFloat(amount), date, type }));
-  };
-
+export default async function IncomePage() {
   return (
     <div>
-      <h1>Create Income</h1>
-      <form onSubmit={handleSubmit}>
-        <input
-          type="number"
-          value={amount}
-          onChange={(e) => setAmount(e.target.value)}
-          placeholder="Amount"
-        />
-        <input
-          type="date"
-          value={date}
-          onChange={(e) => setDate(e.target.value)}
-          placeholder="Date"
-        />
-        <input
-          type="text"
-          value={type}
-          onChange={(e) => setType(e.target.value)}
-          placeholder="Type"
-        />
-        <button type="submit">Submit</button>
-      </form>
+      <h1>Income Management</h1>
+      <IncomeForm />
+      <IncomeList initialData={data} />
     </div>
   );
 }
